move PollutionRenderer class to pollution package
