+++ src/main/java/gregtech/common/GTClient.java
@@ -102,7 +102,7 @@ import gregtech.common.render.GTRendererBlock;
 import gregtech.common.render.LaserRenderer;
 import gregtech.common.render.MetaGeneratedToolRenderer;
 import gregtech.common.render.MultiTileRenderer;
-import gregtech.common.render.PollutionRenderer;
+import gregtech.common.pollution.PollutionRenderer;
 import gregtech.common.render.WormholeRenderer;
 import gregtech.common.render.items.DataStickRenderer;
 import gregtech.common.render.items.InfiniteSprayCanRenderer;
