+++ src/main/java/gregtech/common/pollution/Pollution.java
@@ -45,7 +45,6 @@ import gregtech.api.interfaces.tileentity.IGregTechTileEntity;
 import gregtech.api.net.GTPacketPollution;
 import gregtech.api.util.GTChunkAssociatedData;
 import gregtech.api.util.GTUtility;
-import gregtech.common.render.PollutionRenderer;
 
 public class Pollution {
